Reworked Planet Warps (PoI) plugin (hexicube)
This is a small rework on Hexicube's Planet Warps plugin.
Admins can create Planets of Interest (PoI) by issuing /set_poi <name>
command on a planet, or use /del_poi <name> to removed a PoI
All users can use /poi to show list of available PoI's on the server,
and /poi <name> can be used (only from a ship!) to instantly move to
that PoI

PoI's are intented to be like server hubs, where players can meet
instantly and free of charge.